Add files via upload
Airline reservation System is a computerized system used to store and retrieve information and conduct transactions related to air travel. The project is aimed at exposing the relevance and importance of Airline Reservation Systems. It is projected towards enhancing the relationship between customers and airline agencies using ARSs, and thereby making it convenient for the customers to book the flights as when they require such that they can utilize this software to make reservations.